requirements.txt
pyjwt[crypto]
fastapi
openai
uvicorn[standard]
requests
python-dotenv